writing controller unit tests + refactoring of controller's get and delete methods

+++ src/main/java/ca/sandbox/springrestapitraining/controller/ItemController.java
@@ -13,7 +13,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
+import ca.sandbox.springrestapitraining.exception.MyResourceNotFoundException;
 import ca.sandbox.springrestapitraining.model.Item;
 import ca.sandbox.springrestapitraining.service.ItemService;
 
@@ -33,6 +35,7 @@ public class ItemController {
 	@GetMapping(path = "/{id}")
 	@ResponseStatus(HttpStatus.OK)
 	public ResponseEntity<Item> findOne(@PathVariable Long id) {
+		
 		Optional<Item> item = service.findItem(id);
 		if(item.isEmpty()) {
 			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -43,13 +46,33 @@ public class ItemController {
 	@PostMapping
 	@ResponseStatus(HttpStatus.CREATED)
 	public ResponseEntity<Item> addItem(@RequestBody Item item) {
+		
 		return ResponseEntity.status(HttpStatus.CREATED).body(service.addItem(item));
 	}
 	
 	@DeleteMapping(path = "/{id}")
 	@ResponseStatus(HttpStatus.OK)
 	public void remove(@PathVariable Long id) {
-		service.removeItem(id);
-	}
 
+		try {
+            final Optional<Item> resourceById = checkFound(service.findItem(id));
+            if(resourceById.isPresent()) {
+                service.removeItem(id);
+            } else {
+            	throw new MyResourceNotFoundException();
+            }
+        }
+        catch (MyResourceNotFoundException exc) {
+            throw new ResponseStatusException(
+                HttpStatus.NOT_FOUND, "Item Not Found, can not be deleted", exc);
+        }
+	}
+	
+	public <T> Optional<Item> checkFound(final Optional<Item> optional) {
+        if (optional == null) {
+            throw new MyResourceNotFoundException();
+        }
+        return optional;
+    }
+	
 }
